producers/requirements.txt
confluent-kafka[avro]==1.9.2
numpy==1.23.5
pandas==1.5.3
requests==2.31.0
python-dotenv==1.0.0
